+++ src/firmware/pru_bridge.h
@@ -15,8 +15,8 @@
 #include <stdint.h>
 
 int check_init(void);
-int read_buffer(int ring_no,uint8_t* pru_data,uint8_t length);
-int write_buffer(int ring_no,uint8_t* pru_data,uint8_t length);
+int read_buffer(int ring_no,uint8_t* pru_data,int length);
+int write_buffer(int ring_no,uint8_t* pru_data,int length);
 int check_index(int ring_no);
 
 #endif
